requirements.txt
arrow==1.3.0
asgiref==3.9.1
click==8.2.1
colorama==0.4.6
croniter==6.0.0
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-rq==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
rq==2.5.0
rq-win==0.4.2
six==1.17.0
sqlparse==0.5.3
times==0.7
types-python-dateutil==2.9.0.20250809
tzdata==2025.2
whitenoise==6.9.0
